Show link not connected to storage provider info
Show link not connected to storage provider error when the user is
accessing a link that is not connected to any storage provider. In this
patch, the way to render the container is refactored to a new function
called renderContainer to streamline the logic instead of using tons of
ternary operators. The Icon is also resized to a bigger size.