Fix bug connectedStorageProviders not updated
The bug appears when the user load to the create / edit links after
connecting to storage provider. The issue is that storage providers
options are not updated due to the shared object (user in panel/index.js
shared using UserContext) is not refreshed. In this patch, the
Promise.all is used to bootstrap both the reloadLinks and
fetchUserConnectedStorageProviders calls, making them concurrent and
reduce the waiting time.